Attribute interface completed, some testing needed, see log Patches applied:
 * csardi@szaffi/igraph--main--0.2--patch-108
   Attribute interface completed, some testing needed, see log

The igraph_add_vertices and igraph_add_edges functions now take
the attributes of the newly added nodes/edges as parameters as well.

This has been tested with igraph_read_graph_ncol and the R
attribute handling functions, it works.

The other way (ie. saving the attributes in a file) is not tested yet.
Also some pieces are missing from the R attribute funcions:
- handling of arbitrary data types in add_vertices and add_edges
- attr parameters in add.vertices and add.edges
- handling the vertex and edge selectors in the query functions
- handling of arbitrary data types in the query functions