Replace C versions of ARPACK, LAPACK and BLAS functions with Fortran versions. The non-symmetric ARPACK solver does not work properly with the C versions.
ARPACK is quite picky and works best with the LAPACK and BLAS version
included in its distribution, so we just use these, even if this is
unfortunate, as we might need to include some LAPACK functions twice.
In R this means including them three times, as they are also included
in R itself. :(

This breaks the MSVC compilation, as it is now, but ARPACK can be
compiled on windows separately and the used from igraph as a DLL. 
See http://people.csail.mit.edu/barbic/arpack.html for details.