Fix warning due to factors in graph_from_data_frame
When converting a data.frame with an POSIXct column to an igraph object
using the function 'graph_from_data_frame', a warning is issued due to
ineffective checking of classes:
> Warning message:
> In if (class(newval) == "factor") { :
>   the condition has length > 1 and only the first element will be used

Th warning is true as a POSIXct object has actually two classes:
"POSIXct" and "POSIXt".

To remove the warning, the class of edge and vertex attributes is now
checked using the infix notation "%in%".

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>