Add a parameter loop.size that governs the size of the loops in plot.igraph
Add a parameter loop.size that governs the size of the loops in plot.igraph.

The default value of the parameter is 1, which produces the same behavior as the previous version. I also added the appropriate entry in the help section (new parameter). Larger values produce larger loops, while smaller then 1 produce smaller loops. This is important for me, since the current loops are too big for bigger graphs.

I'm fine with GPL 2 or later and FDL.