Vertex access speedups (#197)
* Speed up `as_adj_list` if igraph_opt("return.vs.es") is false

Update documentation to point out that this option exists.

* Change condition in simple_vs_index to allow short-circuit

This prevent any(is.na(vs)) from getting evaluated if na.ok = TRUE

* Add unsafe_create_vs

This is intended to be useful when the vertex indices
are guaranteed to be in the set of possible vertices.
This is true particularly after C-code when domain
checks are clearly unnecessary. Also allows vertices
to be pre-computed. This speeds up as_adj_list and
ego

* undo change to simple_vs_index

Co-authored-by: Vincent Traag <vincent@traag.net>
Co-authored-by: Tamás Nepusz <ntamas@gmail.com>